Extracted out a part of the player representation class to an abstraction
